requirements.txt
# Core Google AI library for accessing Gemini models
google-generativeai
python-dotenv
google-adk